+++ platform/mac/framework/test_AGDnsProxyListenerFailsInit.mm
@@ -35,6 +35,7 @@ static int regressTestListenerFailsInit() {
                                                 enableDNSSECOK:NO
                                   enableRetransmissionHandling:NO
                                            enableRouteResolver:NO
+                                                      blockEch:NO
                                                     helperPath:nil];
 
     auto *handler = [AGDnsProxyEvents new];