+++ platform/mac/framework/test_AGDnsProxyStandalone.mm
@@ -40,6 +40,7 @@ int main() {
                                                 enableDNSSECOK:NO
                                   enableRetransmissionHandling:NO
                                            enableRouteResolver:NO
+                                                      blockEch:NO
                                                     helperPath:nil];
 
     auto *handler = [AGDnsProxyEvents new];